+++ src/MicroElements.Metadata.AspNetCore/AspNetCore/DependencyInjectionExtensions.cs
@@ -40,10 +40,14 @@ namespace MicroElements.Metadata.AspNetCore
             services.Configure<MetadataJsonSerializationOptions>(configureMetadataJson);
 
             // Allows to serialize property containers
+            // NOTE: Not the same as AspNet JsonSerializerOptions. Use AspNetJsonSerializerOptions wrapper
             services.Configure<JsonSerializerOptions>(options => options.ConfigureJsonForPropertyContainers(configureMetadataJson));
 
             // Configures serialization for AspNetCore
             services.ConfigureJsonOptionsForAspNetCore(options => options.ConfigureJsonForPropertyContainers(configureMetadataJson));
+
+            // Register wrapper for AspNetCore json options.
+            services.AddTransient<AspNetJsonSerializerOptions>(provider => new AspNetJsonSerializerOptions(provider.GetJsonSerializerOptionsOrDefault()));
 
             // Allows to use property containers in swagger
             services.ConfigureSwaggerForPropertyContainers(configureSwaggerOptions);
@@ -102,5 +106,25 @@ namespace MicroElements.Metadata.AspNetCore
 
             return services;
         }
+    }
+
+    /// <summary>
+    /// AspNetCore MVC JsonOptions.Value wrapper that can be used in netstandard.
+    /// </summary>
+    public class AspNetJsonSerializerOptions
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AspNetJsonSerializerOptions"/> class.
+        /// </summary>
+        /// <param name="value"><see cref="JsonSerializerOptions"/> from AspNet host.</param>
+        public AspNetJsonSerializerOptions(JsonSerializerOptions value)
+        {
+            Value = value;
+        }
+
+        /// <summary>
+        /// Gets the <see cref="JsonSerializerOptions"/> from AspNet host.
+        /// </summary>
+        public JsonSerializerOptions Value { get; }
     }
 }