+++ src/MicroElements.Metadata/Metadata/Schema/SimpleTypeSchema.cs
@@ -13,7 +13,7 @@ namespace MicroElements.Metadata.Schema
     /// Example: Currency (string type with maxLength: 3).
     /// </summary>
     [DebuggerTypeProxy(typeof(MetadataProviderDebugView))]
-    public class SimpleTypeSchema : ISchema, IManualMetadataProvider
+    public class SimpleTypeSchema : ISchema, IManualMetadataProvider, ISchemaBuilder<SimpleTypeSchema, ISchemaDescription>
     {
         /// <inheritdoc />
         public string Name { get; }
@@ -24,13 +24,17 @@ namespace MicroElements.Metadata.Schema
         /// <inheritdoc />
         public string? Description { get; }
 
+        /// <inheritdoc />
+        public IPropertyContainer Metadata { get; }
+
         /// <summary>
         /// Initializes a new instance of the <see cref="SimpleTypeSchema"/> class.
         /// </summary>
         /// <param name="name">Type name.</param>
         /// <param name="type">Base type for schema.</param>
         /// <param name="description">Optional schema description.</param>
-        public SimpleTypeSchema(string name, Type type, string? description = null)
+        /// <param name="metadata">Optional metadata for schema.</param>
+        public SimpleTypeSchema(string name, Type type, string? description = null, IPropertyContainer? metadata = null)
         {
             name.AssertArgumentNotNull(nameof(name));
             type.AssertArgumentNotNull(nameof(type));
@@ -39,11 +43,15 @@ namespace MicroElements.Metadata.Schema
             Type = type;
             Description = description;
 
-            Metadata = new ConcurrentMutablePropertyContainer(searchOptions: MetadataProvider.DefaultSearchOptions);
+            Metadata = new ConcurrentMutablePropertyContainer(metadata, searchOptions: MetadataProvider.DefaultSearchOptions);
         }
 
         /// <inheritdoc />
-        public IPropertyContainer Metadata { get; }
+        public SimpleTypeSchema With(ISchemaDescription schemaDescription)
+        {
+            var source = this;
+            return new SimpleTypeSchema(source.Name, source.Type, schemaDescription.Description, source.Metadata);
+        }
     }
 
     /// <summary>
@@ -51,8 +59,9 @@ namespace MicroElements.Metadata.Schema
     /// Simple type does not have own properties but has some restrictions for vase type.
     /// Example: Currency (string type with maxLength: 3).
     /// </summary>
+    /// <typeparam name="T">Schema value type.</typeparam>
     [DebuggerTypeProxy(typeof(MetadataProviderDebugView))]
-    public class SimpleTypeSchema<T> : ISchema<T>
+    public class SimpleTypeSchema<T> : ISchema<T>, ISchemaBuilder<SimpleTypeSchema<T>, ISchemaDescription>
     {
         /// <inheritdoc />
         public Type Type => typeof(T);
@@ -75,17 +84,12 @@ namespace MicroElements.Metadata.Schema
             Name = name;
             Description = description;
         }
-    }
-
-    public class NullTypeSchema : ISchema
-    {
-        /// <inheritdoc />
-        public string Name => "null";
 
         /// <inheritdoc />
-        public Type Type => typeof(object);
-
-        /// <inheritdoc />
-        public string? Description => "Null type.";
+        public SimpleTypeSchema<T> With(ISchemaDescription schemaDescription)
+        {
+            var source = this;
+            return new SimpleTypeSchema<T>(source.Name, schemaDescription.Description).SetMetadataFrom(source);
+        }
     }
 }